% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.pendensity-package}
\alias{mlr3learners.pendensity}
\alias{mlr3learners.pendensity-package}
\title{mlr3learners.pendensity: Learners from the {pendensity} package for 'mlr3'}
\description{
Adds learners from the {pendensity} package to
    'mlr3'.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk}

}
